post completion commit

+++ src/main/java/com/example/help_desk/repository/AuditLogRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.help_desk.entity.AuditLogs;
 
-public interface AuditLogRepository extends JpaRepository<AuditLogs, Integer> {
+public interface AuditLogRepository extends JpaRepository<AuditLogs, Long> {
 }
